FEAT(product-page): Add loader for product page, move user to header, when load new data
Jira: [[FE]Add loader for product page](https://cozyhometeam.atlassian.net/browse/CH-756)
