+++ src/pages/ProductPage/ProductPage.scss
@@ -1,4 +1,5 @@
 @use '../../style/variables.scss' as *;
+@import '../../style/mixins.scss';
 
 body {
 	position: relative;
@@ -6,6 +7,7 @@ body {
 
 .product-page {
 	&__wrapper {
+        position: relative;
 		display: flex;
 		justify-content: space-between;
 		gap: 34px;
@@ -60,6 +62,10 @@ body {
 
     &__interested {
         margin-top: 130px;
+    }
+
+    &__loader {
+        @include loader-blur;
     }
 }
