+++ src/shared-components/Loader/Loader.scss
@@ -3,7 +3,6 @@
     align-items: center;
     justify-content: center;
     user-select: none;
-    pointer-events: none;
     &_image {
         width: 70px;
         height: 70px;
