+++ src/store/reducers/cartSlice.ts
@@ -53,19 +53,20 @@ export const fetchProductCartInfo = createAsyncThunk(
         controller = new AbortController();
 
         try {
-            const response = await fetch(`${API_BASE}product/basket`, {
-                method: 'POST',
-                body: JSON.stringify([...cartBody]),
-                headers: {
-                    'Content-type': 'application/json; charset=UTF-8',
-                },
-                signal: controller.signal,
-            });
+            // const response = await fetch(`${API_BASE}product/basket`, {
+            //     method: 'POST',
+            //     body: JSON.stringify([...cartBody]),
+            //     headers: {
+            //         'Content-type': 'application/json; charset=UTF-8',
+            //     },
+            //     signal: controller.signal,
+            // });
 
-            const result = await response.json();
-            if (!response.ok) throw new Error('something went wrong');
+            // const result = await response.json();
+            // if (!response.ok) throw new Error('something went wrong');
 
-            return result;
+            // return result;
+            return [];
         } catch (error: any) {
             if (error.name === 'AbortError') {
                 return thunkAPI.rejectWithValue('');
