+++ src/style/mixins.scss
@@ -42,5 +42,4 @@
     background-color: rgba(255, 255, 255, 0.3);
     backdrop-filter: blur(1px);
     z-index: 3;
-    user-select: none;
 }
